Small bug fixes in HRA and UI fixes.

+++ features/hra/src/main/java/com/caressa/hra/common/HraHelper.kt
@@ -48,7 +48,7 @@ object HraHelper {
         heightWeightDialog.show()
     }
 
-    fun showWeightDialog( weight :Int,layWeight: CustomEditTextHra ,listener : HraQuesBmiFragment,context: Context ) {
+    fun showWeightDialog( weight :Double,layWeight: CustomEditTextHra ,listener : HraQuesBmiFragment,context: Context ) {
         val data = ParameterDataModel()
         data.title = "Weight"
         data.value = " - - "
